requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
motor==3.6.0
# pymongo is installed automatically as motor dependency

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12

# Validation & Settings
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.2.0

# CORS & Security
slowapi==0.1.9

# Email (placeholder - choose provider later)
aiosmtplib==3.0.2

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.28.0
